Added rendering functionality to Celeritas
In this update, we've introduced a new feature to the Celeritas package. We've added a Render field to the Celeritas struct and implemented a createRender function. This function initializes a new instance of the Render struct from the render package with necessary configurations. Now, each instance of Celeritas will have its own renderer which can be used for various rendering tasks.